requirements.txt
cached-property==1.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
h5py==3.7.0
idna==3.3
imageio==2.19.2
kiwisolver==1.4.2
kornia==0.6.4
matplotlib==3.5.2
munkres==1.1.4
numpy==1.22.4
onnx==1.11.0
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.1.1
pip==22.1
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyparsing==3.0.9
PyQt5==5.15.4
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scipy==1.8.1
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==62.2.0
shortuuid==1.0.9p
sip==6.5.1
six==1.16.0
smmap==5.0.0
tensorboardX==2.5.1
thop==0.1.0.post2206102148
toml==0.10.2
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.64.0
typing_extensions==4.2.0
unicodedata2==14.0.0
urllib3==1.26.9
Wand==0.6.7
wandb==0.12.16
wheel==0.37.1
